requirements.txt
Flask==3.1.0
python-telegram-bot==21.7
